Broke out all informational functions
Broke out all informational functions into file informationModule.js.
